+++ packages/plugins/GeoLocation/src/language.ts
@@ -15,6 +15,8 @@ const language: LanguageOption[] = [
           },
           toast: {
             notInBoundary: 'Sie befinden sich nicht im Kartengebiet.',
+            boundaryError:
+              'Die Überprüfung Ihrer Koordinate ist fehlgeschlagen. Bitte versuchen Sie es später erneut oder wenden Sie sich an einen Administrator, wenn das Problem bestehen bleibt.',
           },
         },
       },
@@ -34,6 +36,8 @@ const language: LanguageOption[] = [
           },
           toast: {
             notInBoundary: "You are not within the map's boundaries.",
+            boundaryError:
+              'Validating your coordinate failed. Please try later again or contact an administrator if the issue persists.',
           },
         },
       },