+++ packages/plugins/Pins/CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## unpublished
+
+- Feature: Updating `@polar/lib-passes-boundary-check` to major version 2. (TODO: Publish and install library before publishing this update in a new version.)
+- Feature: Adding `boundaryOnError` parameter to let user define behaviour on boundary check errors.
+
 ## 1.0.0
 
 Initial release.