+++ packages/types/custom/CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## unpublished
+
+- Feature: Add parameter `boundaryOnError` to `PinsConfiguration`.
+- Feature: Add parameter `boundaryOnError` to `GeoLocationConfiguration`.
+
 ## 1.0.0
 
 Initial release.